# api/services/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Cocktail Bar",
    "description": "Documentation of Coctail Bar API"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "Root route. Use it to check if the API is running.",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/recipes/all": {
      "get": {
        "description": "Get all recipes.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          }
        }
      }
    },
    "/recipes/all/count": {
      "get": {
        "description": "Count all recipes.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          }
        }
      }
    },
    "/recipes/all/list-{prop}": {
      "get": {
        "description": "List all values stored under the specified property.",
        "parameters": [
          {
            "name": "prop",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          },
          "400": {
            "description": "Invalid property name specified in the request. Response body contains error details."
          }
        }
      }
    },
    "/recipes/all/list-{prop}/distinct": {
      "get": {
        "description": "List all distinct values stored under the specified property.",
        "parameters": [
          {
            "name": "prop",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          },
          "400": {
            "description": "Invalid property name specified in the request. Response body contains error details."
          }
        }
      }
    },
    "/recipes/{key}-equals/{value}": {
      "get": {
        "description": "Get all Recipes where the value stored under the specified key equals the specified value (case insensitive).",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "value",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          },
          "400": {
            "description": "Invalid value format specified in the request. Response body contains error details."
          }
        }
      }
    },
    "/recipes/{key}-equals/{value}/list-{prop}": {
      "get": {
        "description": "List all values stored under the specified property in Recipes where the value stored under the specified key equals the specified value (case insensitive).",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "value",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "prop",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          },
          "400": {
            "description": "Invalid property name or value format specified in the request. Response body contains error details."
          }
        }
      }
    },
    "/recipes/{key}-equals/{value}/list-{prop}/distinct": {
      "get": {
        "description": "List all distinct values stored under the specified property in Recipes where the value stored under the specified key equals the specified value (case insensitive).",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "value",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "prop",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          },
          "400": {
            "description": "Invalid property name or value format specified in the request. Response body contains error details."
          }
        }
      }
    },
    "/recipes/{key}-equals/{value}/count": {
      "get": {
        "description": "Count all Recipes where the value stored under the specified key equals the specified value (case insensitive).",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "value",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          },
          "400": {
            "description": "Invalid value format specified in the request. Response body contains error details."
          }
        }
      }
    },
    "/recipes/{key}-contains/{value}": {
      "get": {
        "description": "Get all Recipes where the value stored under the specified key contains the specified value (case insensitive).",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "value",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          },
          "400": {
            "description": "Invalid value format specified in the request. Response body contains error details."
          }
        }
      }
    },
    "/recipes/{key}-contains/{value}/list-{prop}": {
      "get": {
        "description": "List all values stored under the specified property in Recipes where the value stored under the specified key contains the specified value (case insensitive).",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "value",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "prop",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          },
          "400": {
            "description": "Invalid property name or value format specified in the request. Response body contains error details."
          }
        }
      }
    },
    "/recipes/{key}-contains/{value}/list-{prop}/distinct": {
      "get": {
        "description": "List all distinct values stored under the specified property in Recipes where the value stored under the specified key contains the specified value (case insensitive).",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "value",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "prop",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          },
          "400": {
            "description": "Invalid property name or value format specified in the request. Response body contains error details."
          }
        }
      }
    },
    "/recipes/{key}-contains/{value}/count": {
      "get": {
        "description": "Count all Recipes where the value stored under the specified key contains the specified value (case insensitive).",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "value",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          },
          "400": {
            "description": "Invalid value format specified in the request. Response body contains error details."
          }
        }
      }
    },
    "/recipes/update/by-name/{name}": {
      "post": {
        "description": "Update a Recipe with the specified name. Data to update is specified in the request body.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          }
        }
      }
    },
    "/recipes/update/by-id/{id}": {
      "post": {
        "description": "Update a Recipe with the specified id. Data to update is specified in the request body.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          },
          "400": {
            "description": "Invalid id format specified in the request. Response body contains error details."
          }
        }
      }
    },
    "/recipes/create": {
      "post": {
        "description": "Create a new Recipe. Data is specified in the request body.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          }
        }
      }
    },
    "/recipes/duplicate/by-name/{name}": {
      "post": {
        "description": "Duplicate a Recipe with the specified name (case sensitive).",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          }
        }
      }
    },
    "/recipes/duplicate/by-id/{id}": {
      "post": {
        "description": "Duplicate a Recipe with the specified id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          },
          "400": {
            "description": "Invalid id format specified in the request. Response body contains error details."
          }
        }
      }
    },
    "/recipes/delete/by-name/{name}": {
      "post": {
        "description": "Delete a Recipe with the specified name (case sensitive).",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          }
        }
      }
    },
    "/recipes/delete/by-id/{id}": {
      "post": {
        "description": "Delete a Recipe with the specified id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          },
          "400": {
            "description": "Invalid id format specified in the request. Response body contains error details."
          }
        }
      }
    },
    "/recipes/get/by-name/{name}": {
      "get": {
        "description": "Get the Recipe with the specified name (case sensitive).",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          }
        }
      }
    },
    "/recipes/get/by-id/{id}": {
      "get": {
        "description": "Get the Recipe with the specified id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Recipe"
            }
          },
          "500": {
            "description": "Internal Server Error. Response body contains error details."
          },
          "400": {
            "description": "Invalid id format specified in the request. Response body contains error details."
          }
        }
      }
    }
  }
}